query로 옮김
